Clean: Relocate STEX BWPixelProxy
