+++ src/imageproc/BWPixelProxy.h
