+++ src/imageproc/CMakeLists.txt
@@ -7,6 +7,7 @@ SET(
         BinaryThreshold.cpp BinaryThreshold.h
         SlicedHistogram.cpp SlicedHistogram.h
         ByteOrder.h BWColor.h
+        BWPixelProxy.h
         ConnComp.h Connectivity.h
         BitOps.cpp BitOps.h
         SeedFill.cpp SeedFill.h
